Make $B accessible in exception factories
Exception factories are now created by the Function() constructor, which
only has access to global variables. $B is defined with the var keyword
at the top level. The top level context is the global context in the browser,
but it's the module context in node, so exception factories don't by default
have access to $B in node.